Day 15 - part 01: finished solution
* Use JGraphT to build a hallway graph
* Calculate the shortest path from the start location to the oxygen
  location using breadth first-search

+++ src/main/java/nl/smerik/adventofcode/aoc2019/controller/Day15Controller.java
@@ -17,7 +17,7 @@ public class Day15Controller {
     }
 
     @GetMapping("/part-01")
-    public ResponseEntity<Long> getSolutionPart01() {
+    public ResponseEntity<Integer> getSolutionPart01() {
         return ResponseEntity.ok(dayService.getSolutionPart1());
     }
 }
